☕ chore (utils): move everything in a folder and use index.ts
