--- src/lib/string.ts
@@ -1,3 +1,0 @@
-export function toTitleCase(str: string) {
-  return str.replace(/\w\S*/g, (txt) => txt.charAt(0).toUpperCase() + txt.substring(1).toLowerCase());
-}
