+++ src/lib/types.ts
@@ -14,13 +14,4 @@ export interface dicebearOptions {
   style?: string;
 }
 
-export type CookieOptions = {
-  path?: string;
-  httpOnly?: boolean;
-  sameSite?: "strict" | "lax" | "none";
-  secure?: boolean;
-  maxAge?: number;
-  expires?: Date;
-};
-
 export type AttendanceView = "today" | "yesterday" | "history";
